Moved code from Leiden to separate source file.
